+++ app-backend/api/src/main/scala/annotation-project/Routes.scala
@@ -178,6 +178,12 @@ trait AnnotationProjectRoutes
             } ~ delete {
               deleteTask(projectId, taskId)
             }
+          } ~ pathPrefix("status") {
+            pathEndOrSingleSlash {
+              put {
+                updateTaskStatus(projectId, taskId)
+              }
+            }
           } ~ pathPrefix("lock") {
             pathEndOrSingleSlash {
               post {