+++ app-backend/db/src/main/resources/migrations/V78__index_label_class_in_labels_to_classes_join.sql
@@ -1,0 +1,1 @@
+CREATE INDEX IF NOT EXISTS annotation_labels_annotation_label_classes_class_id_idx ON annotation_labels_annotation_label_classes (annotation_class_id);